% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_max_row}
\alias{get_max_row}
\title{Get Row with Maximum Acreage}
\usage{
get_max_row(df)
}
\arguments{
\item{df}{A data frame containing an 'Acreage' column.}
}
\value{
Returns the row from the data frame that has the maximum value in the 'Acreage' column.
}
\description{
This function takes a data frame and returns the row with the maximum acreage.
It assumes that the data frame has a column named 'Acreage'.
}
\examples{
data <- data.frame(Acreage = c(100, 200, 150), Name = c("Farm1", "Farm2", "Farm3"))
get_max_row(data)

}
