% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_top_crops_acerage}
\alias{get_top_crops_acerage}
\title{Get Rows Meeting a Specified Cutoff}
\usage{
get_top_crops_acerage(df, cutoff = 0.9)
}
\arguments{
\item{df}{A data frame containing an 'Acreage' column.}

\item{cutoff}{A numeric value (default is 0.90) specifying the cumulative percentage cutoff.
Rows are selected such that the cumulative acreage is just above this cutoff.}
}
\value{
Returns a subset of the input data frame, including rows up to the point
        where the cumulative percentage of acreage first exceeds the specified cutoff.
}
\description{
This function sorts a data frame by the 'Acreage' column in descending order and
returns the top rows based on a cumulative percentage cutoff of the 'Acreage' column.
It assumes the data frame contains an 'Acreage' column.
}
\examples{
data <- data.frame(Acreage = c(100, 200, 150, 50), Name = c("Farm1", "Farm2", "Farm3", "Farm4"))
get_top_rows(data, 0.90)

}
